+++ Kernel/Bus/USB/UHCI/UHCIController.cpp
@@ -6,7 +6,7 @@
  */
 
 #include <AK/Platform.h>
-#include <Kernel/Bus/USB/UHCIController.h>
+#include <Kernel/Bus/USB/UHCI/UHCIController.h>
 #include <Kernel/Bus/USB/USBRequest.h>
 #include <Kernel/CommandLine.h>
 #include <Kernel/Debug.h>
