+++ Kernel/Bus/USB/USBPipe.cpp
@@ -5,7 +5,7 @@
  */
 
 #include <Kernel/Bus/USB/PacketTypes.h>
-#include <Kernel/Bus/USB/UHCIController.h>
+#include <Kernel/Bus/USB/UHCI/UHCIController.h>
 #include <Kernel/Bus/USB/USBPipe.h>
 #include <Kernel/Bus/USB/USBTransfer.h>
